+++ src/Stormancer.Plugins/Users/Stormancer.Server.Plugins.Users/Changelog.rst
@@ -17,6 +17,7 @@ Changed
 - Updated dependency to Users to support new abstraction lib.
 - By default, data are not stored in any database anymore
 - Use Stormancer.Server.Plugins.Users.EntityFramework to use entity framework to store data.
+- Analytics now store realtime session metrics grouped per dimensions vector.
 
 Added
 *****